requirements.txt
beautifulsoup4==4.8.2
certifi==2019.11.28
chardet==3.0.4
entrypoints==0.3
flake8==3.7.9
idna==2.9
mccabe==0.6.1
pycodestyle==2.5.0
pyflakes==2.1.1
pymongo==3.10.1
python-dotenv==0.12.0
requests==2.23.0
schedule==0.6.0
soupsieve==2.0
urllib3==1.25.8
wget==3.2
